requirements.txt
asyncio
discord.py
sqlalchemy